+++ lib/index.ts
@@ -1,2 +1,2 @@
-export * from './static-website-deployment';
-export * from './static-website-pipeline';
+export * from "./deployment";
+export * from "./pipeline";